Initial version of configurable shaders
