+++ pixels.py
@@ -1,4 +1,5 @@
 from commands import Sleep
+from shaders.pixel_shaders import PressedShader
 
 class PixelListener:
     BRIGHTNESS = 0.3
@@ -75,35 +76,7 @@ class PixelListener:
         self.sleeping = False
 
     def highlight(self, key_index):
-        self.shaders += [DefaultShader(self, key_index)]
+        self.shaders += [PressedShader(self.pixels, key_index)]
 
     def reset(self, key_index):
         pass
-
-class DefaultShader:
-    MAX_FRAMES = 4
-    listener = None
-    key_index = None
-    key_index_max = 0
-    frame_index = None
-
-    def __init__(self, listener, key_index):
-        self.listener = listener
-        self.key_index = key_index
-        self.frame_index = DefaultShader.MAX_FRAMES
-        self.key_index_max = min(len(self.listener.keycolors), PixelListener.MAX_LEDS)
-
-    def tick(self, _, frames):
-        self.frame_index -= frames
-        if self.key_index >= self.key_index_max: return
-
-        if self.frame_index > 0:
-            color_val =  0xFF * (self.frame_index / DefaultShader.MAX_FRAMES)
-            self.listener.pixels[self.key_index] = (color_val, color_val, color_val)
-            self.listener.pixels.show()
-        else:
-            self.listener.pixels[self.key_index] = self.listener.keycolors[self.key_index]
-            self.listener.pixels.show()
-        
-    def done(self):
-        return self.frame_index <= 0
